bugfix
1、添加格式化文档；
2、新功能bugfix；（lasso生成gef缺失max值；get_multilabel_regiondata_bgef偶现crash）

+++ cgefCellgem.cpp
@@ -1032,7 +1032,7 @@ void cgefCellgem::writeCell_new()
                     ptr->m_cx,
                     ptr->m_cy,
                     offset,
-                    ptr->m_map_gExp.size(),
+                    (unsigned short)ptr->m_map_gExp.size(),
                     ptr->m_expcnt,
                     ptr->m_dnbcnt,
                     ptr->m_area,
